#1080 fix Double->BigDecimal tests and relate services.
git-svn-id: svn://svn.opentaps.org/opentaps_all/branches/ofbiz-0904-1.4.0-preview1@13122 d3523486-f5fe-0310-a41a-a15a4e76f3c7

+++ hot-deploy/opentaps-common/src/common/org/opentaps/domain/inventory/InventoryServiceInterface.java
@@ -66,7 +66,7 @@ public interface InventoryServiceInterface extends ServiceInterface {
      * Sets the transfer quantity, required parameter for {@link #prepareInventoryTransfer}.
      * @param xferQty a <code>Double</code> value
      */
-    public void setXferQty(Double xferQty);
+    public void setXferQty(BigDecimal xferQty);
 
     /**
      * Sets the inventory transfer identifier to cancel.
